Circular heatmap

+++ Utilities/src/main/java/nl/harmjanwestra/utilities/graphics/panels/Panel.java
@@ -33,7 +33,7 @@ public abstract class Panel {
 		this.height = panelHeight;
 	}
 
-	public void drawRotate(Graphics2D g2d, double x, double y, int angle, String text) {
+	public void drawRotate(Graphics2D g2d, double x, double y, double angle, String text) {
 		g2d.translate((float) x, (float) y);
 		g2d.rotate(Math.toRadians(angle));
 		g2d.drawString(text, 0, 0);
